Adding a completed manual flow to a hunt. (#1182)
It is now possible to add an unrelated flow to a hunt. This helps in
the case where the original flow in the hunt timed out. The user then
can re-run the hunt manually possibly increasing timeout. Then they
can simply click the add flow to hunt button in the UI to add the flow
to an existing time.

Additionally, the upload function accepts all 4 timestamps for a
file. Depending on the uploader these may be used or just recorded. In
the case of the file uploader, the filesystem timestamps will be reset
to reflect these times if possible. This allows us to create the
Windows.KapeFiles.Extract artifact to extract the container files
recreating timestamps.