Display the hostname in the host overview
Previously the UI only showed the Fqdn and not the hostname. The
hostname is reported by the machine itself (on windows using wmi). The
FQDN is sometimes the same and sometimes different since it does a dns
lookup.

The two terms are separate but related so we should show both.