Compiling client event table ignores missing artifacts
If an artifact is added to the client monitoring table and then
deleted the client monitoring service will fail to load because it is
unable to recompile the missing artifact.

This PR makes it a non-fatal error, the event table will simply skip
the missing artifact.

Also some small fixes:
1. USN parser returns the time.Time object
2. UI shows JSON mode in raw json views