Artifact search optimization
When searching for artifacts by name we used to return the entire
definition which would result in over 1.6mb or response json. For this
reason total number of hits was limited to 100 artifacts.

This change returns only the names of matching artifacts which makes
the response much smaller and therefore we can return up to 1000
artifacts in the search box.