Created Server.Internal.Interrogation artifact (#1306)
Previously, when a hunt targets by OS, and a new client appears, the
interrogation service races the hunt service to check the client's
hunt membership. Usually the hunt manager will examine the client
immediately and since we have no information about its OS we can not
assign it to the hunt.

This PR adds a watcher to the hunt manager to check for new
interrogated clients, since these may change their OS property. In
order to determine when a client is updated this adds another internal
Server.Internal.Interrogation artifact which is guaranteed to
fire **after** the interrogation is processed.