Default proxy is now settable via config file.
Previously the only way to override the default http proxy was through
the http_proxy environment variable. This PR also adds the ability to
specify the proxy in the config file:

Client:
  proxy: http://127.0.0.1:3128/
  server_urls:
  - https://192.168.1.11:8000/

Or for Frontend:

Frontend:
  proxy: http://127.0.0.1:3128/

Proxy setting can still be overriden with environment variables as
well but this makes it easier to set the default.