Refactored queue listener to preserve order. (#1340)
The QueueManager implements a mediator between consumers of events and
producers of events. This mediator is called the Listener object. The
main goal of the Listener is to disconnect producers and consumers -
so if the consumer is too slow, the producer automatically queues
events into the buffer file so as not to block the producer.

In the previous implementation it was possible for events to be
reordered as events were diverted to the file buffer.

This PR fixes this behavior:
1. When an event arrives at the Listener, the Listener tried to
   forward it directly to the listener channel.
2. If this is not possible because the listener is currently busy,
   the code switches to file buffering mode. The event is pushed to the
   file buffer and all further events received will go into the file
   buffer.
3. We pump events from the file buffer to the lister channel in
   another goroutine. When the file buffer is fully drained, we switch
   back into direct delivery mode.

This preserves ordering of events and ensures that if the readers are
quick enough all events will be delivered directly bypassing the file
buffer.