BUGFIX: Datastore on windows is unable to represent files with .
On Windows filenames that end with "." are not created correctly
because windows assumes the "." is an extension and strips it
off. This causes corruption in the datastore because generally "." are
not santized in filenames but in this special case they may not appear
at the end of the filename.

The bug manifested in indexing issues where the index was unable to
represent certain hostnames (where the length was exactly right to
make the partition directory end with "."). Therefore the offline
collector was unable to find the existing client and regenerated
client ids for the same hostname.

This PR fixes this and also updates the index search to remove
indexing bare terms - now all indexed terms contain an action
version (host: label: etc).