Notify currently connected clients on new hunts
When a hunt is started we used to notify all the clients, but this
creates unnecessary churn on the system. In this PR we specifically
forward all connected clients to the hunt manager instead of notifying
all clients. This keeps clients connected and avoids a thundering herd
condition.

The PR also adds completion function to QueueMessageForClient() so it
can be notified after the message is queued.