Client info manager now keeps track of scheduled tasks.
Previously the minion would perform a datastore ListChildren()
operations each time a client connected to ensure there are no
outstanding tasks. Now the fact that no tasks are present can be
cached in the client info cache allowing the minion to return an empty
response without making a datastore operation.

This change reduced the total numebr of datastore operations leading
to better throughput.