Initial artifacts can be specified in config
When the server first initializes, these server artifacts will be
collected. Given this mechanism it is possible to configure the server
to perform arbitrary initialization tasks (e.g. preload artifact
exchange etc).

Also added a `scope` accessor which operates similarly to the `data`
accessor except the filename specifies a scope variable. This is
useful for chaining into accessors which require a url because url
encoding is not binary safe.