Fixed bug in client comms.
When the client received a 503 status from the server (due to load
shedding for example) it failed to close the socket on error
paths. This resulted in a leak of a socket and presented in having
many TIME_WAIT sockets.

This PR refactors the http comms code to ensure the socket is closed
on all error paths