Tag artifacts as built ins (#1467)
Previously we used the name of the artifact to determine if it was
built in or not. This forces a naming convention on the
artifacts (must start with Custom) and it was quite difficult to
manage.

The whole reason for the naming is to prevent users from overwriting
built in artifacts. This PR tags artifacts as being built in and so
can prevent users overwriting them without needing a particular naming
convention.

Fixed: #1466