add limitations to description and key path to query (#1514)
Added a limitations section as this artifact parses the live registry. If a user is not logged in their NTUSER will not be parsed. Similarly there is no current support for parsing VSS/Regback stored data.

Also included the keypath to the query to allow examiners to quickly identify which value the key was located in (for ex user and software hive).