Refractor of accessors and path manipulations (#1545)
Filesystems are inherently directed graphs (usually trees). An abstract "path" is a list of components specifying a path from the root to traverse. When we call an OS API we need to provide a string to the OS we term an "OS Path".

One way to think about paths is that the OS path is a serialisation of an abstract "OSPath". Different operating systems have a different way of actually serialising the same abstract OSPath.

For example, a the following components based abstract path ["C:", "Windows", "System32"] will be serialised by windows into C:\\Windows\\System32 and by Linux conventions as /C:/Windows/System32

This refactor converts all path handling internally into an abstract representation which get serialised by the various accessors. This allows us to deal with paths in an OS agnostic way.

This change also refactors the accessors to separate them in the code base.

Also introduced the concept of a device manager which handles accessor registration in a specified way - this allows accessors to be remapped for dead disk forensic applications.