Make GitHub actions more flexible on Windows (#1549)
There are a number of pathnames that depend on the $GITHUB_WORKSPACE
path being d:\a\velociraptor\velociraptor.  This works for the main
project repository but breaks when the repository uses a different name.

This commit uses drive substitution to map $GITHUB_WORKSPACE to a new X:
drive that can be used directly.