Merged FileSystemAccessorFactory with FileSystemAccessor.
When remapping accessors we build a template accessor and remap into
the scope's DeviceManager. However remapped accessors need to have the
scope's context from before the remapping. For example, if the "from"
accessor needs to open another accessor, it should be before the
remapping.

Therefore we now capture a copy of the global device managed into a
subscope when preparing the remapping accessor - it will be able to
use that to open the pristine accessors before remapping.