Added rekey() VQL function (#1604)
This forces the client to regenerate it's own client id and reconnect
using the new client id.

Sometimes client writeback files are accidentally copied across
deployments (e.g. in VM images). This PR adds a
Server.Monitor.ClientConflict server event artifact to watch for the
409 conflict events and force clients to rekey when detected. It
should be used in situations where it is possible to client writeback
files to be duplicated.

This PR also fixes the client comms mechanism so it can be gracefully
cancelled and recreated without restarting the process.