Error collection based on VQL logs (#1672)
Previously it was difficult to know when a collection actually failed
because the VQL might have completed correctly but it did not achieve
what it was set out to. Despite this the Velociraptor server marked
the collection as successful.

This PR adds a new log regex setting which can be used to trigger
collection failure when the log is detected.

Also refactored to allow further collection results to be accepted
even after the collection is failed. Added further tests for this.

Fixed hunt stat collection which keeps track of how many collection
are errored.