Update auto accessor to re-open file with ntfs if read failed
The auto accessor is used to automatically switch to the ntfs accessor
when we can not open a file due to it being locked but some locking
mechanisms allow the file to be properly opened but then fail the read
operation.

This PR wraps the file accessor and re-opens the file with the ntfs
accessor if the file fails to read.