GUI Table widgets now can apply transformations on the table. (#1740)
When users use the GUI much of the time they need to manipulate the
table in a simple way. Currently users have to export to csv and
import to excel so they can filter or sort - or they can use the
notebook to apply VQL to do the filtering and sorting (which has a
learning curve).

This PR makes the GUI table able to filter and sort by itself. It
works by creating the transformation on the original result set and
storing it separately.

This way all functionality works on the transformed result
set (e.g. exporting to csv or json).