artifacts: add artifactset parameter type
Historically, the artifacts that handle monitoring of client events for
forwarding to other services have been somewhat clunky to configure.
Originally, it involved an integrated hidden CSV file and a series of
bools that referenced fields within that CSV.  More recently, at least
one has been converted to use a visible CSV that at least gives access
to the CSV editing interface in the GUI.  It's still not a great
interface.

This feature queries the API for a list of artifacts that match the
specified artifact type (or all of them if unspecified) and uses the
list to create a UI consisting of a group of switches (with names and
tooltips describing each one) that can be used to select the artifacts.
It then uses those to assemble the CSV format parameter that the
artifacts already expect so existing configurations will continue to
work without updating.  Since it's CSV behind the scenes, the default
definitions will be consumed as before as well.

To use this interface, an Artifact must be defined using:
type: artifactset

and optionally:
artifact_type: {SERVER,SERVER_EVENT,CLIENT,CLIENT_EVENT,etc}