Bugfix: Server monitoring queries were not correctly cancelled.
Server monitoring tables were previously insensitive to changes in
artifact definitions. In 0.6.4 an event watcher was added to allow the
tables to be updated when an artifact was modified.

This triggered a bug where old queries were not properly cancelled,
causing multiple queries of the same kind to be present (e.g. in issue
https://github.com/Velocidex/velociraptor/issues/1764)

This change fixes that issue and adds a test. We also ensure that the
server event table is not reset when the queries have not
changed (e.g. an unrelated artifact was modified or the artifact was
not modified significantly).