Bugfix: VFS viewer was unable to access files with \ in name
On Linux the path separator is '/' and '\' is allowed in side file
names. Within VQL this is handled properly using the OSPath
abstraction - all paths a just a list of components regardless of
the path separator.

The VFS GUI however was still creating collections for
System.VFS.DownloadFile and System.VFS.ListDirectory with a serialized
string path. This leads to VQL having to parse the string back into an
OSPath object which can create confusion in edge cases.

This PR sidesteps the entire issue by extending these artifacts to
accept a list of components instead of a serialized path. The GUI can
then maintain this list and launch the artifacts with the components
already split.

This PR also adds the ability to pass a list of components to any
OSPath VQL parameter which will just build the right OSPath object
from these.