Added a default process tracker
If no process tracker event monitoring rule is used, the process
tracker is set to the default one. The default tracker is equivalent
to a `pslist()` for each access.

Existing artifacts may access the process tracker instead of calling
pslist() for enrichments. Then if the user elects to have an active
tracker installed all existing artifacts will benefit automatically.