Added orgs() plugin and user management (#1949)
It is not possible in VQL to add users to different orgs and grant
them different access.

Also allow the query() plugin to runas a user and run within a
different org context. This allows scheduling of collection hunts
across different orgs as well retrieving the results from different
orgs.