collect each query's status separately (#2049)
When dealing with multiple queries per collection it is important to
keep stats on each query separately to get more accurate picture of
what happened. For example one artifact may collect successfuly and
another may fail.

This change also fixes a bug where the execution duration was
sometimes counted twice for queries that failed because they send an
error as well as a completion message.