Sync 0.6.6 branch to master  (#2054)
* [Snyk] Upgrade ace-builds from 1.9.3 to 1.9.5 (#2037)

fix: upgrade ace-builds from 1.9.3 to 1.9.5

Snyk has created this PR to upgrade ace-builds from 1.9.3 to 1.9.5.

See this package in npm:
https://www.npmjs.com/package/ace-builds

See this project in Snyk:
https://app.snyk.io/org/scudette/project/76f4d127-566b-42ef-86f4-bdcbc92b90b4?utm_source=github&utm_medium=referral&page=upgrade-pr

* Added Send to CyberChef context menu on table cells. (#2039)

* Made context menus settable in the config file (#2040)

Added more context targets:
- VirusTotal
- Google
- CyberChef

* Updated themes for context menu (#2041)

Added "Clipboard" target

* Added org_delete() function to remove orgs. (#2042)

* Org manager maintains services lifetime (#2045)

* Pass org ids in href parameters (#2047)

Usually OrgId is passed in headers but we do have some href links so
we need to also support passing the org id in query parameters.

* collect each query's status separately (#2049)

When dealing with multiple queries per collection it is important to
keep stats on each query separately to get more accurate picture of
what happened. For example one artifact may collect successfuly and
another may fail.

This change also fixes a bug where the execution duration was
sometimes counted twice for queries that failed because they send an
error as well as a completion message.

* Split server sanity checks into root org and other orgs (#2052)

The sanity checker ensured the server is running in a sane state. It
also starts things like initial server artifacts etc.

This PR split sanity checks to run on the root org (i.e. once when the
server is started) and on each org. For example the initial server
artifacts are only run on the root org and not on other orgs.

Added org rm command

* Fixed CSS for column selector ui (#2053)

* Prepare for 0.6.6 release

Co-authored-by: Snyk bot <snyk-bot@snyk.io>