Allow collect_client() etc to accept ArtifactSpec protobuf (#2067)
When inspecting Flow objects we often get an ArtifactContext protobug
with contains an ArtifactSpec. This is different from the dict() based
spec we use in VQL so it is difficult to feed the output of
e.g. hunt_flows() into collect_client()

This PR makes it so collect_client() can handle the ArtifactSpec type
transparently by converting it to a dict based spec.