auth/github: report github errors when encountered during callback
I encountered an issue when updating to 0.6.6 where I did not specify
a public_url and GitHub authentication would fail immediately.  From the
GUI, it silently returned to the root page.  In the log it only reported
that the response was missing the access_token variable.

It turns out that GitHub returns errors in its response querystring that
can be reported to the user to aid in diagnosing failures.  This commit
checks to see if an error is present before proceeding with processing
the token.