Uploaded files are now shows with client paths (#2116)
The uploaded files table in the uploaded files tab was showing the
file store path for the uploaded files. This is not that relevant to
the user who is more interested in the filesystem path on the client
end.

This change allows the filestore path and the path shown to be
different.