Provide the option to expand sparse files in export (#2124)
Sparse files are collected by Velociraptor in a compact way with an
index file ensuring accurate reconstruction. However when exporting
for parsing by expernal tools we typically need the file to be padded
out.

This PR allows the export to be optionally padded.

NOTE: This can lead to huge files! use with care.