Implemented a max memory grouper
Previous GROUP BY operations were done in memory. This PR allows
Velociraptor to fall back to a disk based algorithm when there are too
many bins (i.e. the GROUP BY cadinality is too high).

The disk based algorithm is about 2-3 times slower than the in memory
one but ensures we do not use unlimited memory. The threashold at
which to switch to the disk based algorithm is determined by the
config file.