Add Public Key Encryption Support to Offline Collections (#2133)
This PR adds the `pk_encrypt` and `pk_decrypt` functions for encrypting and decrypting data using public key encryption. It also modifies the collector and container code to allow for embedding of metadata in both encrypted and unencrypted archives. This PR also adds the `server_frontend_cert` function for retrieving the frontend's public cert. Finally, it modifies the webui to be able to select through different encryption schemes when creating offline collectors. 

`pk_encrypt` and `pk_decrypt` default to using the server's frontend cert and private key when their `public_key` and `private_key` fields are unspecified. With `pk_decrypt`, you must be a server admin to decrypt data using the private key.


These changes allow for the creation of unique session keys per collection. Those session keys get encrypted using the provided public key, thus ensuring the password used to create the ZIP is not leaked. The encrypted session key goes inside of `metadata.json` in the top level of the zip archive. Using the corresponding private key, this session key can be decrypted and then used to decrypt the zip archive.