Retry reads on EOF in NTFS accessor
The NTFS context gets reset periodically and this causes a read to
return EOF on closed file handle. When this happens we need to retry
the read which will force it to recreate caches and parse MFT etc.

This is usually not needed for single shot collections, so this PR
also disables it from the KapeTargets artifact