Propagate user's prefered timezone for export tables (#2232)
When the user sets their preferred timezone in the user preferences
the UI adjusts the output. This PR also ensures this preference is
propagated to the download table API so the encoded timestamps are
also presented in the correct timezone.

The user may always adjust the timezone back to UTC to receive their
times in UTC again.