Bugfix: Do not expand sparse files when importing
The offline collection zip may contain very sparse files (e.g. $J) so
we do not want to automatically expand those files when importing as a
collection.

This PR introduces a new accessor "collector_sparse" which does not
automatically expand sparse files, instead the importer copies the
index file as it is.