Bugfix: Store client specific dashboard in client space. (#2255)
The old Reports are migrated to the notebook renderer which needs to
write the result to the datastore. This requires choosing an
appropriate location for the notebook on disk. Previously all
notebooks were put in the global notebooks directory but some reports
are specific for each client.

This PR refactores the paths of where reports will be stored.

Currently reports are used for:

1. Home screen (dashboard)  (type: "SERVER_EVENT")
2. Client's VQL Drilldown screen (type: "CLIENT")
3. View Artifacts screen (type: "ARTIFACT_DESCRIPTION")
4. Client Monitoring report (to be deprecated - same as monitoring notebook).