api: add v1/GetUser/{name} endpoint to return user record
To build the user admin GUI, we'll need all of the details for a given
user.  This endpoint will return permission denied if a user requests
the record for another user unless the caller has the SERVER_ADMIN role
and ORG_ADMIN for at least one of the orgs listed in the record.