Bugfix: uploads.json in the flow download refered to filestore paths (#2282)
The _Components column of the uploads.json in the filestore refers to filestore
paths (e.g. clients/C.1bfa6928675831f5-O123/collections/F.CE2PSBS6BQCSO...) This makes it incompatible with the offline collector which uses this column to refer to paths within the zip file itself.

Previously this same file was just copied to the container but now it is being modified to support the different components list. This makes the upload.json format compatible with the offline collector so that VQL artifacts designed to work on the offline collector can also work on flow downloads.