Fixed Windows.Sys.Users artifact.
On domain controllers the users() VQL plugin will dump the entire AD
but this is rarely what we want! We usually just want to find local
users only.

Previously the artifact materialized the entire AD and filtered using
an O(n^2) algorithm which means it often timed out on domain
controllers.

The VQL is now:
- Limited to 1000 users on a DC
- Use group by to deduplicate the results from the registry scan

** Also updated title attributes to use the tooltip CSS so they are
   rendered more clearly.