Use pageable tables for the VFS (#2343)
The Virtual File system was storing the entire directory listing in
the data store which made it very inefficient for large
directories. This PR stores a reference into the VFSListDirectory
collection so we only store a very small file in the data store. This
also allows us to use the regular Table browser to view the results,
making it possible to handle directories of arbitrary size.