utils/time.js: fix handling of nanosecond-resolution timestamps
The logging infrastructure produces timestamps with nanosecond
resolution but the javascript spec for the Date() API only handles
up to millisecond resolution (as well as a different format for timezones).

This results in the log table showing "Invalid date" under Timestamp and
causes an uncaught exception in the tooltip that shows relative time.

This commit strips out the additional digits and reformats the time
zone to the expected format:
2022-08-19 19:56:29.261590674 +0000 UTC becomes 2022-08-19 19:56:29.261+00:00