Refactor client flow context manager
Now FlowStats are sent for all queries even after they are
teminated. This makes it much easier to track which query terminated,
and what the status is. The ArtifactCollectorContext is now made up
from two parts:

1. A list of QueryStatus items one for each query (including stats for
   each query and final status code).
2. Derived fields from the above which represent the collection wide
   totals.

Added more tests.

Server Artifact service is now refactored to use the same
FlowRequest/FlowStats messages as the client.