Implemented a preview Column renderer
Sometimes it is handy to present data inside the VQL generated
table. Previously the only way to do this was to include the data in
the output JSON using e.g. read_file().

This approach is problematic because the JSON size is limited by the
GUI. For larger data we really want to be able to upload the entire
file using the VQL upload() function.

The `preview_upload` column renderer takes the output of the upload()
function and renders an GUI widget in the table that allows closer
inspection of the collected file. It also automatically renders images
inside the table cell if the uploaded file was an image (otherwise a
snippet of the raw data).

This approach allows us to keep the JSON lean while allowing for large
quantity of data to be delivered for inspection by analysts.