Uploads are now deduplicated on store_as_name. (#2415)
It is often convenient in VQL to repeat the same finding from the same
file on multiple rows. But previously each invocation of the upload()
function would cause the file to be uploaded again

Although it is possible to work around this using the cache() function
it is more useful to only upload the same file once but provide a
reference to the upload on each row.

This PR deduplicates the uploads by store_as_name.