Added a tempfile based materializer to have safe queries (#2420)
Sometimes people materialize queries which are very large eg.

LET X <= SELECT * FROM glob(...)

The glob may be huge! Previously this query was stored in memory
causing issues for large queries. This PR adds a tempfile based
materializer which automatically serializes large LET into a file,
then reads the results from the file. This allows the above construct
to be used safely with an upper bound on memory use.