Store client path components in the uploads metadata
This makes it a lot easier to use them later when creating the
downloads export. Previously the download code had to derive the
client's original components from the filestore components but this is
not always possible to do reliably. Carrying the original client's
path components through makes it more reliable and easier.