Refactored repack functionality into a VQL function (#2461)
Previously repacking (as used by the offline collector) was
implemented in VQL by calling the main binary with the repack
command. This is problematic on servers which remove the execve()
functionality.

This PR implements the repack functionality in VQL and therefore makes
it available for use in restricted environments. It also simplifies
the CreateCollector artifact.

Also added a Server.Utils.CreateMSI artifact which automatically
prepares an MSI for use within an org.