Hex column types did not required hex encoding
Hex columns were encoding the string value in the column. Generally
this is incorrect as the string valu comes from JSON and if the VQL
query emits raw binary data into the JSON then it will be corrupted
while contecting to UTF8.

This PR requires a column declared as "hex" to be actually hex encoded
similarly to the base64 columns. This helps to correct the VQL because
non-encoded data will simply not show up. This is preferable to almost
correct data which has UTF8 errors in it.