Fixed Test
Raw Registry accessor repeated the names of keys that have values of
the same name (i.e. default value for a key name). This go dudplicated
leading to loss of the value name. This fix returns a RawRegValueInfo
object which also behaves as a key (looks like a directory) so it can
be recursed into but also can be read as a Value (file).

Also this PR refactores the server artifact runner so it does not
require scheduling a task but it gets launched immediately.