Deduplicate glob hits (#2490)
Where multiple glob expressions were used and the same file was found
by multiple expressions the path was not properly deduplicated.

Also improved locking on the upload deduplication. When one thread is
uploading, other threads should wait on it to complete.

Raw Registry accessor repeated the names of keys that have values of
the same name (i.e. default value for a key name). This go dudplicated
leading to loss of the value name. This fix returns a RawRegValueInfo
object which also behaves as a key (looks like a directory) so it can
be recursed into but also can be read as a Value (file).

Also this PR refactores the server artifact runner so it does not
require scheduling a task but it gets launched immediately.