Added a max_row_buffer_size parameter
Previously we controlled the amount of buffering on the client using
max_rows and max_wait. Limiting by number of rows is not fine grained
enough because the size of each row depends on how much data is in
it.

This PR introduces a max_row_buffer_size parameter which flushes the
client's row buffer when this size is reached. This ensures we do not
exceed message sizes which helps keep the POST size under the
specified limits.