Create monitoring_logs.go (#2611)
I wanted a way to check the overall health of client event monitoring
within a notebook (or, later, artifacts in a dashboard), similar to how
we can aggregate events from various clients using the source plugin. It
looked like the only existing access to monitoring_log data was in the
tables code that powers the Client Events page in the gui. It seemed
easier and safer to make a new plugin instead of overloading the
source() plugin in the results package.

In local tests, I got the expected results for both server and client
events, but I'm sure there are things to refactor or fix here. :-)

<img width="1569" alt="Screen Shot 2023-04-05 at 11 54 09 PM"
src="https://user-images.githubusercontent.com/40576678/230270086-1a9907f6-939f-4934-bd89-8cdee7a41744.png">

---------

Co-authored-by: Mike Cohen <mike@velocidex.com>