Added a lockdown mode to the server config.
When a server is locked down certain permissions are removed (even
from admins). The lockdown is set in the config file. This helps to
mitigate the risk of a Velociraptor server admin account compromise.

After initial deployment, the administrator can set the server in
lockdown by adding the following config directive:

lockdown: true

to the server.config.yaml. After this the following permissions will
be denied:

 - ARTIFACT_WRITER
 - SERVER_ARTIFACT_WRITER
 - EXECVE
 - SERVER_ADMIN
 - FILESYSTEM_WRITE
 - FILESYSTEM_READ
 - MACHINE_STATE

During an active IR the server may be taken out of lockdown by
removing the directive from the config file and restarting the
service. Usually the config file is only writable by root and the
Velociraptor server is running as a low privilege account which can
not write to the config file.